function decodeTimestamp (buf, size, headerSize) moved to helper
