+++ lib/decoder.js
@@ -1,7 +1,7 @@
 'use strict'
 
 var bl = require('bl')
-var { IncompleteBufferError } = require('./helpers.js')
+var { IncompleteBufferError, decodeTimestamp } = require('./helpers.js')
 
 const SIZES = { 0xc4: 2, 0xc5: 3, 0xc6: 5, 0xc7: 3, 0xc8: 4, 0xc9: 6, 0xca: 5, 0xcb: 9, 0xcc: 2, 0xcd: 3, 0xce: 5, 0xcf: 9, 0xd0: 2, 0xd1: 3, 0xd2: 5, 0xd3: 9, 0xd4: 3, 0xd5: 4, 0xd6: 6, 0xd7: 10, 0xd8: 18, 0xd9: 2, 0xda: 3, 0xdb: 5, 0xde: 3, 0xdc: 3, 0xdd: 5 }
 
@@ -224,33 +224,6 @@ module.exports = function buildDecode (decodingTypes) {
 
     var value = decode(toDecode)
     return buildDecodeResult(value, headerSize + size)
-  }
-
-  function decodeTimestamp (buf, size, headerSize) {
-    var seconds
-    var nanoseconds = 0
-
-    switch (size) {
-      case 4:
-      // timestamp 32 stores the number of seconds that have elapsed since 1970-01-01 00:00:00 UTC in an 32-bit unsigned integer
-        seconds = buf.readUInt32BE(0)
-        break
-
-      case 8:
-      // Timestamp 64 stores the number of seconds and nanoseconds that have elapsed
-      // since 1970-01-01 00:00:00 UTC in 32-bit unsigned integers, split 30/34 bits
-        var upper = buf.readUInt32BE(0)
-        var lower = buf.readUInt32BE(4)
-        nanoseconds = upper / 4
-        seconds = ((upper & 0x03) * Math.pow(2, 32)) + lower // If we use bitwise operators, we get truncated to 32bits
-        break
-
-      case 12:
-        throw new Error('timestamp 96 is not yet implemented')
-    }
-
-    var millis = (seconds * 1000) + Math.round(nanoseconds / 1E6)
-    return buildDecodeResult(new Date(millis), size + headerSize)
   }
 }
